backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langgraph==0.0.46
langchain==0.1.20
langchain-openai==0.1.6
pydantic==2.5.0
httpx==0.25.2
redis==5.0.1
psycopg2-binary==2.9.7
pytrends==4.9.2
numpy==1.24.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
tavily-python==0.3.3
beautifulsoup4==4.12.3
spacy==3.7.4
# python -m spacy download en_core_web_sm
